requirements.txt
asgiref==3.11.1
Django==6.0.3
django-filter==25.2
django-rest-registration==0.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
sqlparse==0.5.5
whitenoise==6.12.0
